Implementing savegame()
Testing with an added  fxml file and controller. Not the actual game

+++ MotelMurders/src/Acquaintance/IData.java
@@ -6,7 +6,7 @@ public interface IData {
     
     //Methods for data highscore - Make methods as abstract/generalized as possible f.eks. Collection instead of arraylist
 	Object LoadHighscore();
-        void SaveHighscore(int seconds, String playerName, int MAXTIME);
+        void SaveHighscore(double seconds, String playerName, int MAXTIME);
         void SaveGame(Object[] game);
         Collection<Object>LoadGame();
